+++ R/gggenes-package.R
@@ -1,0 +1,7 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom lifecycle deprecate_warn
+## usethis namespace: end
+NULL